name correction and restructuring
